requirements.txt

langchain-community>=0.4.0
langchain-core>=0.3.0
langgraph>=0.2.0
neo4j>=6.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
